Organize the code into tests and src folders
