+++ pytest.ini
@@ -1,4 +1,4 @@
 # content of pytest.ini
-#[pytest]
-#testpaths = tests
-#python_paths = ./src/
+[pytest]
+testpaths = tests
+python_paths = ./src/
